Updated flags with respect to deleted files.
